+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.2 (November 6, 2018)
+
+CHANGES:
+
+* Addition of new script to list all RDS instances and details about them. ([@TGWolf][])
+* Tidy up of variable and function names to give them more meaning. ([@TGWolf][])
+* Added 'exit code' mode - sets the exit code to be the number of real upgrades available. ([@TGWolf][])
+* Added 'silent' mode so no output it given (designed to work with exit code above). ([@TGWolf][])
+
 ## 1.0.1 (November 6, 2018)
 
 CHANGES: